Minor change in graph construction
Removed the 80% query coverage requirement for connecting sequences with an edge. While the script performs all pairwise alignments, it does not do so in both directions (if A is aligned to B, but B is not aligned to A). Percent identity is commutative (the direction of the alignment doesn't matter), but query coverage is not (it depends on which sequence is the query), so including that requirement without performing every pairwise alignment may miss some homologies. An alternative would have been to change lines 105 and 106
from:
    for j in range(len(fileList)):
        for i in range(j+1):
to:

    for j in range(len(fileList)):
        for i in range(len(fileList)):

But this would take twice as long to run.